# math: replace exported variables with runtime.KeepAlive in benchmarks